Reorganise test helpers
