+++ event-store/src/lib.rs
@@ -14,6 +14,8 @@ pub mod event_context;
 pub mod event_saver;
 pub mod pg;
 pub mod store_query;
+#[doc(hidden)]
+pub mod test_helpers;
 
 pub use crate::aggregator::*;
 pub use crate::amqp::*;
@@ -21,6 +23,8 @@ pub use crate::event::Event;
 pub use crate::event_saver::*;
 pub use crate::pg::*;
 pub use crate::store_query::*;
+#[doc(hidden)]
+pub use crate::test_helpers::*;
 
 #[cfg(test)]
 mod tests {
